requirements.txt
## This are mist.io dependencies with loose version restrictions.
## Keeping this file helps discover what the actual dependencies are and
## assists in the process of upgrading them. Occasionally we build using this
## requirements file, test everything exhaustively, and then store the output
## of `pip freeze` to requirements-frozen.txt. Using frozen/pinned requirements
## ensures that the build won't break because of a new release of some
## dependency.

amqp<3.0
apscheduler
beautifulsoup4
boto3
dnspython
dateparser
dramatiq
elasticsearch
elasticsearch_tornado
flake8
future
gevent
greenlet
ipaddress # requests needs this for dns matching
ipython
iso8601
jsonpatch
jsonpickle
kombu
mongoengine
mongomock
names
netaddr
parse
passlib
pastedeploy  # Required for uwsgi paste logging, can probably be removed.
pastescript  # Required for uwsgi paste logging, can probably be removed.
pika
pingparsing
pretty
pycryptodome
pylibmc
pymongo
pyyaml
pyramid
pyramid_chameleon
pytest
python3-openid
python-gitlab
pyvmomi
requests
s3cmd
scp
sendgrid-python
sentry-dramatiq
sentry-sdk
sockjs-tornado
tornado
troposphere
#tornado_profile
uwsgidecorators
websocket-client